refactor: move the canvas card state and snapshot hooks out of page.tsx
Refs #1936.

Canvas page: 4,753 lines before, 4,467 lines after.
